refactor: consolidate frontend authentication and skeleton components
- Merge useAuth and useCanvasAuth hooks into single implementation
- Update all imports to use consolidated auth hook
- Consolidate LoadingSkeleton and skeleton.tsx into enhanced LoadingSkeleton
- Remove duplicate skeleton.tsx file
- Update LoadingSkeletonProps to extend Chakra UI SkeletonProps
- Add forwardRef support and additional props to LoadingSkeleton
