+++ frontend/src/components/Common/LoadingSkeleton.tsx
@@ -1,25 +1,27 @@
 import type { LoadingSkeletonProps } from "@/types/components"
-import { Skeleton, VStack } from "@chakra-ui/react"
-import { memo } from "react"
+import { Skeleton, Stack } from "@chakra-ui/react"
+import { forwardRef } from "react"
 
-export const LoadingSkeleton = memo(function LoadingSkeleton({
-  height = "20px",
-  width = "100%",
-  lines = 1,
-}: LoadingSkeletonProps) {
-  if (lines === 1) {
-    return <Skeleton height={height} width={width} />
-  }
+export const LoadingSkeleton = forwardRef<HTMLDivElement, LoadingSkeletonProps>(
+  function LoadingSkeleton(
+    { height = "20px", width = "100%", lines = 1, gap = 2, ...rest },
+    ref,
+  ) {
+    if (lines === 1) {
+      return <Skeleton height={height} width={width} ref={ref} {...rest} />
+    }
 
-  return (
-    <VStack gap={2} align="stretch">
-      {Array.from({ length: lines }).map((_, index) => (
-        <Skeleton
-          key={index}
-          height={height}
-          width={index === lines - 1 ? "80%" : width}
-        />
-      ))}
-    </VStack>
-  )
-})
+    return (
+      <Stack gap={gap} width="full" ref={ref}>
+        {Array.from({ length: lines }).map((_, index) => (
+          <Skeleton
+            key={index}
+            height={height}
+            width={index === lines - 1 ? "80%" : width}
+            {...rest}
+          />
+        ))}
+      </Stack>
+    )
+  },
+)
