+++ frontend/src/components/UserSettings/DeleteConfirmation.tsx
@@ -16,7 +16,7 @@ import {
   DialogTrigger,
 } from "@/components/ui/dialog"
 import { useCustomToast, useErrorHandler } from "@/hooks/common"
-import useAuth from "@/hooks/useCanvasAuth"
+import { useAuth } from "@/hooks/auth"
 
 const DeleteConfirmation = () => {
   const [isOpen, setIsOpen] = useState(false)
